image: color glyph support for image backend
This patch enables color glyph support for image backend.  The patch depends
on the following 3 patches:

1. 00991ac6 HACK Disable use of pixman glyph API
2. 5f9bb65 HACK Route xlib glyph compositing through fallback path to
get color emo
3. fc57558 WIP Adjust image compositor for color emoji bitmap glyphs

The only problem I see is that when there is glyph overlapping, it still
uses ADD operator to add glyph images on a mask surface.  This means
not only RGB values are added, but also the alpha value.  This can throw
off the color because FreeType rendered color glyph is premultiplied